requirements.txt
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
charset-normalizer==3.3.2
db-service-grpc @ git+https://github.com/alexseipopov/capybaras_db_handler_grpc.git@main
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
jmespath==1.0.1
pillow==10.2.0
protobuf==4.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
storage-service-grpc @ git+https://github.com/CapyDevelop/storage_service_grpc.git@main
urllib3==2.0.7
